Developed the mcu_terminal.py CLI application. Will test the windows implementation/module soon. Works fine on Linux// PART 2: changes to root directory now included

+++ PWR_Control_F401RCT6/Core/Inc/comms.h
@@ -14,5 +14,5 @@ void Comms_SendHeartbeat(void);
 typedef void (*Comms_OnHandshake_t)(uint16_t telemetry_ms, uint8_t send_ack);
 void Comms_RegisterHandshakeCb(Comms_OnHandshake_t cb);
 
-extern volatile uint32_t heartbeat_counter;
+extern volatile uint16_t heartbeat_counter;
